backend/requirements.txt
Flask
opencv-python
numpy
torch
transformers
pillow